apps/models/pyproject.toml
[project]
name = "models"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "bitsandbytes>=0.45.5 ; sys_platform != 'darwin'",
    "datasets>=3.5.0",
    "fastapi[standard]>=0.112.1",
    "huggingface-hub>=0.30.2",
    "joblib>=1.4.2",
    "keybert>=0.9.0",
    "peft>=0.15.2",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "python-whois>=0.9.5",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "tldextract>=5.2.0",
    "transformers>=4.51.3",
    "triton>=3.2.0 ; sys_platform != 'darwin'",
    "validators>=0.34.0",
]
